Changed price type to double.

+++ database/migrations/2016_02_22_173945_add_status_and_price_to_rooms_table.php
@@ -5,31 +5,19 @@ use Illuminate\Database\Migrations\Migration;
 
 class AddStatusAndPriceToRoomsTable extends Migration {
 
-	/**
-	 * Run the migrations.
-	 *
-	 * @return void
-	 */
 	public function up()
 	{
 		Schema::table('rooms', function(Blueprint $table)
 		{
-			//
 			$table->string('availability');
-			$table->integer('price');
+			$table->double('price');
 		});
 	}
 
-	/**
-	 * Reverse the migrations.
-	 *
-	 * @return void
-	 */
 	public function down()
 	{
 		Schema::table('rooms', function(Blueprint $table)
 		{
-			//
 			$table->dropColumn('availability');
 			$table->dropColumn('price');
 		});
